pre-review refactorings
