--- modules/ml/src/test/java/org/apache/ignite/ml/tree/boosting/GDBRegressionOnTreesTrainerTest.java
@@ -1,7 +1,0 @@
-package org.apache.ignite.ml.tree.boosting;
-
-import static org.junit.Assert.*;
-
-public class GDBRegressionOnTreesTrainerTest {
-
-}
