+++ modules/ml/src/test/java/org/apache/ignite/ml/tree/randomforest/RandomForestClassifierTrainerTest.java
@@ -21,6 +21,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import org.apache.ignite.ml.composition.ModelOnFeaturesSubspace;
 import org.apache.ignite.ml.composition.ModelsComposition;
 import org.apache.ignite.ml.composition.predictionsaggregator.OnMajorityPredictionsAggregator;
 import org.apache.ignite.ml.tree.DecisionTreeConditionalNode;
@@ -68,6 +69,10 @@ public class RandomForestClassifierTrainerTest {
 
         RandomForestClassifierTrainer trainer = new RandomForestClassifierTrainer(4, 3, 5, 0.3, 4, 0.1);
         ModelsComposition model = trainer.fit(sample, parts, (k, v) -> k, (k, v) -> v);
+        model.getModels().forEach(m -> {
+            assertTrue(m instanceof ModelOnFeaturesSubspace);
+            assertTrue(((ModelOnFeaturesSubspace) m).getMdl() instanceof DecisionTreeConditionalNode);
+        });
 
         assertTrue(model.getPredictionsAggregator() instanceof OnMajorityPredictionsAggregator);
         assertEquals(5, model.getModels().size());
